requirements.txt
line-bot-sdk
flask~=1.1.2
python-dotenv~=0.15.0
gunicorn
requests~=2.25.1
pytest
coverage